SBERDOMA-1028 moved delete button with modal into Division page
